refactor: 計算ロジックをCoreプロジェクトとして分離する。
