+++ settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'youkai-password-challenge'
 
-include 'youkai-front', 'youkai-server'
-include 'youkai-console'
+include 'youkai-front', 'youkai-core', 'youkai-server'
+include 'youkai-core', 'youkai-console'
