+++ youkai-console/build.gradle
@@ -24,13 +24,10 @@ jar {
     }
 }
 
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.2'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.2'
-}
-
 test {
     useJUnitPlatform()
 }
 
+dependencies {
+    implementation project(':youkai-core')
+}
