+++ youkai-core/.gitignore
@@ -1,0 +1,23 @@
+.gradle
+/build/
+/bin/
+!gradle/wrapper/gradle-wrapper.jar
+/*.log
+*.swp
+
+### STS ###
+.apt_generated
+.classpath
+.factorypath
+.project
+.settings
+.springBeans
+.sts4-cache
+
+### IntelliJ IDEA ###
+.idea
+*.iws
+*.iml
+*.ipr
+/out/
+/src/main/generated/
