+++ youkai-core/build.gradle
@@ -1,0 +1,17 @@
+plugins {
+    id 'application'
+    id 'java'
+}
+
+sourceCompatibility = versions.sourceCompatibility
+
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.2'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.2'
+}
+
+test {
+    useJUnitPlatform()
+}
+
