+++ youkai-server/build.gradle
@@ -14,6 +14,8 @@ configurations {
 }
 
 dependencies {
+    implementation project(':youkai-core')
+
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
     implementation 'org.springframework.boot:spring-boot-starter-jdbc'
